fix : bug : delete advisoe for DualProjectController

+++ database/migrations/2025_09_09_134519_update_dual_project_reports_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
 
             $table->boolean('is_concluded')->default(false);
             $table->boolean('is_hired')->default(false);
-            $table->integer('qualification')->nullable();
+            $table->decimal('qualification', 5, 2)->nullable();
             $table->enum('advisor', ['interno', 'externo'])->default('interno');
         });
     }
